feat: change js to ts -update dependencies -change js to ts -make scripts for package.json
