+++ package.json
@@ -2,12 +2,13 @@
   "name": "portfolio",
   "version": "1.0.0",
   "description": "Benjamin Catouillard's Portfolio",
-  "main": "server.js",
+  "main": "dist/src/server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build:sass": "node-sass -r --output-style compressed public/sass/style.scss -o public/css",
+    "build:sass": "node-sass -r --output-style compressed public/sass/style.scss -o dist/public/css",
     "watch:sass": "npm run build:sass && npm run build:sass -- -w",
-    "start": "npx nodemon server.js"
+    "copy:assets": "copyfiles public/**/*.css public/**/*.webp public/**/*.ico public/**/*.js public/*.txt public/*.xml dist/",
+    "copy:views": "copyfiles ./src/views/*.pug ./src/views/**/*.pug dist/",
+    "start": "tsc && node dist/src/server.js"
   },
   "repository": {
     "type": "git",
@@ -26,12 +27,19 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "helmet": "^4.4.1",
-    "mysql": "^2.18.1",
     "node-fetch": "^2.6.1",
-    "node-sass": "^4.14.1",
+    "node-sass": "^6.0.1",
     "nodemailer": "^6.5.0",
+    "nodemon": "^2.0.12",
     "pug": "^3.0.2",
-    "serve-favicon": "^2.5.0",
-    "webpack": "^4.46.0"
+    "serve-favicon": "^2.5.0"
+  },
+  "devDependencies": {
+    "@types/compression": "^1.7.2",
+    "@types/express": "^4.17.13",
+    "@types/serve-favicon": "^2.5.3",
+    "copyfiles": "^2.4.1",
+    "tslint": "^6.1.3",
+    "typescript": "^4.4.3"
   }
 }
