+++ src/views/default.pug
@@ -3,7 +3,7 @@ html(lang="fr")
   head
     title Benjamin Catouillard - #{title}
     block css
-      link(rel='stylesheet', href='/css/style.css')
+      link(rel='stylesheet', href='css/style.css')
     meta(name="viewport" content="width=device-width, initial-scale=1.0 minimum-scale=1.0, maximum-scale=5.0, user-scalable=yes")
     meta(name="Content-Type" content="text/html; charset=utf-8")
     meta(name="description" content=description)
